refs is now an INT32 again.
Rev: src/array.h:1.24
Rev: src/constants.h:1.13
Rev: src/gc.h:1.59
Rev: src/interpret.h:1.63
Rev: src/mapping.h:1.30
Rev: src/multiset.h:1.15
Rev: src/object.h:1.52
Rev: src/pike_types.h:1.41
Rev: src/program.h:1.99
Rev: src/stralloc.h:1.48

+++ src/array.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: array.h,v 1.23 2000/08/03 17:55:50 grubba Exp $
+ * $Id: array.h,v 1.24 2000/08/09 12:49:27 grubba Exp $
  */
 #ifndef ARRAY_H
 #define ARRAY_H
@@ -14,7 +14,7 @@
 
 struct array
 {
-  ptrdiff_t refs;		/* Reference count */
+  INT32 refs;		/* Reference count */
 #ifdef PIKE_SECURITY
   struct object *prot;
 #endif
